Fixed 2 bugs & added some commenting
Bug #1: All of the methods that gave the overstaffed, understaffed & unmanaged shifts were not displaying properly. This was due to a break statement on line 133 of Roster.java, and has been replaced with a continue.
Bug #2: Formatting of getRosterForDay was incorrect - fixed using substring to truncate the end of the string off in line 64 of Shift.java